--- src/main/java/gov/usgs/wma/wqp/openapi/annotation/ResultCountResponse.java
@@ -1,24 +1,0 @@
-package gov.usgs.wma.wqp.openapi.annotation;
-
-import static java.lang.annotation.ElementType.METHOD;
-import static java.lang.annotation.ElementType.PARAMETER;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import gov.usgs.wma.wqp.openapi.model.ResultCountJson;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-
-@Retention(RUNTIME)
-@Target({ METHOD, PARAMETER })
-@ApiResponse(
-		responseCode = "200",
-		description = "OK",
-		content = @Content(schema = @Schema(implementation = ResultCountJson.class))
-		)
-public @interface ResultCountResponse {
-
-}
