--- src/main/java/gov/usgs/wma/wqp/openapi/annotation/post/CountStationPostOperation.java
@@ -1,30 +1,0 @@
-package gov.usgs.wma.wqp.openapi.annotation.post;
-
-import static java.lang.annotation.ElementType.METHOD;
-import static java.lang.annotation.ElementType.PARAMETER;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import gov.usgs.wma.wqp.openapi.model.StationCountJson;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-
-@Retention(RUNTIME)
-@Target({ METHOD, PARAMETER })
-@Operation(
-		description = "Return anticipated record counts. Use when the list of parameter values is too long for a query string."
-				+ " See the HEAD endpoint for valid values.",
-		responses = {
-				@ApiResponse(
-						responseCode = "200",
-						description = "OK",
-						content = @Content(schema = @Schema(implementation = StationCountJson.class)))
-				}
-		)
-public @interface CountStationPostOperation {
-
-}
