+++ src/main/java/gov/usgs/wma/wqp/openapi/model/post/BasePostParams.java
@@ -2,7 +2,7 @@ package gov.usgs.wma.wqp.openapi.model.post;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class BasePostParams {
+public abstract class BasePostParams {
 
 	@JsonProperty("mimeType")
 	String mimeType;
