+++ src/main/java/gov/usgs/wma/wqp/openapi/model/post/OrganizationPost.java
@@ -1,0 +1,8 @@
+package gov.usgs.wma.wqp.openapi.model.post;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class OrganizationPost extends BasePostParams {
+	@JsonProperty("organization")
+	public String[] organization;
+}
