+++ src/main/java/gov/usgs/wma/wqp/openapi/model/post/SummaryMonitoringLocationPost.java
@@ -2,7 +2,7 @@ package gov.usgs.wma.wqp.openapi.model.post;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class SummaryOrganizationPost extends OrganizationPost {
+public class SummaryMonitoringLocationPost extends MonitoringLocationPost {
 	@JsonProperty("summaryYears")
 	public String summaryYears;
 }
