+++ src/main/java/gov/usgs/wma/wqp/webservice/ProjectController.java
@@ -11,11 +11,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.MediaType;
 import org.springframework.web.accept.ContentNegotiationStrategy;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -24,22 +22,18 @@ import gov.usgs.wma.wqp.dao.intfc.IStreamingDao;
 import gov.usgs.wma.wqp.mapping.Profile;
 import gov.usgs.wma.wqp.mapping.delimited.ProjectDelimited;
 import gov.usgs.wma.wqp.mapping.xml.IXmlMapping;
-import gov.usgs.wma.wqp.parameter.FilterParameters;
-import gov.usgs.wma.wqp.service.ConfigurationService;
-import gov.usgs.wma.wqp.service.ILogService;
 import gov.usgs.wma.wqp.openapi.ConfigOpenApi;
-import gov.usgs.wma.wqp.openapi.annotation.ActivityCountResponse;
+import gov.usgs.wma.wqp.openapi.annotation.FormUrlPostOperation;
 import gov.usgs.wma.wqp.openapi.annotation.GetOperation;
 import gov.usgs.wma.wqp.openapi.annotation.HeadOperation;
 import gov.usgs.wma.wqp.openapi.annotation.PostCountOperation;
 import gov.usgs.wma.wqp.openapi.annotation.PostOperation;
 import gov.usgs.wma.wqp.openapi.annotation.query.FullParameterList;
-import gov.usgs.wma.wqp.openapi.annotation.query.MimeTypeJson;
-import gov.usgs.wma.wqp.openapi.annotation.query.MimeTypeStd;
-import gov.usgs.wma.wqp.openapi.annotation.query.Zip;
 import gov.usgs.wma.wqp.openapi.model.ProjectCountJson;
+import gov.usgs.wma.wqp.parameter.FilterParameters;
+import gov.usgs.wma.wqp.service.ConfigurationService;
+import gov.usgs.wma.wqp.service.ILogService;
 import gov.usgs.wma.wqp.util.HttpConstants;
-import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -68,8 +62,6 @@ public class ProjectController extends BaseController {
 
 	@HeadOperation
 	@FullParameterList
-	@MimeTypeStd
-	@RequestMapping(method=RequestMethod.HEAD)
 	public void projectHeadRequest(
 			HttpServletRequest request,
 			HttpServletResponse response,
@@ -80,8 +72,6 @@ public class ProjectController extends BaseController {
 
 	@GetOperation
 	@FullParameterList
-	@MimeTypeStd
-	@GetMapping()
 	public void projectGetRequest(
 			HttpServletRequest request,
 			HttpServletResponse response,
@@ -91,40 +81,33 @@ public class ProjectController extends BaseController {
 	}
 
 	@PostOperation
-	@MimeTypeStd
-	@Zip
-	@PostMapping(consumes=MediaType.APPLICATION_JSON_VALUE)
 	public void projectJsonPostRequest(
 			HttpServletRequest request,
 			HttpServletResponse response,
 			@Parameter(hidden=true) @RequestParam(value="mimeType", required=false) String mimeType,
 			@Parameter(hidden=true) @RequestParam(value="zip", required=false) String zip,
-			@RequestBody FilterParameters filter
+			@Parameter(hidden=true) @RequestBody FilterParameters filter
 			) {
 		doDataRequest(request, response, filter, mimeType, zip);
 	}
 
-	@Operation(description="Same as the JSON consumer, but hidden from swagger", hidden=true)
+	@FormUrlPostOperation
 	@PostMapping(consumes=MediaType.APPLICATION_FORM_URLENCODED_VALUE)
 	public void activityFormUrlencodedPostRequest(HttpServletRequest request, HttpServletResponse response, FilterParameters filter) {
 		doDataRequest(request, response, filter);
 	}
 
 	@PostCountOperation
-	@ActivityCountResponse
-	@MimeTypeJson
-	@Zip
 	@ApiResponse(
 			responseCode="200",
 			description="OK",
 			content=@Content(schema=@Schema(implementation=ProjectCountJson.class)))
-	@PostMapping(value="count", consumes=MediaType.APPLICATION_JSON_VALUE, produces=MediaType.APPLICATION_JSON_VALUE)
 	public Map<String, String> projectPostCountRequest(
 			HttpServletRequest request,
 			HttpServletResponse response,
 			@Parameter(hidden=true) @RequestParam(value="mimeType", required=false) String mimeType,
 			@Parameter(hidden=true) @RequestParam(value="zip", required=false) String zip,
-			@RequestBody FilterParameters filter
+			@Parameter(hidden=true) @RequestBody FilterParameters filter
 			) {
 		return doPostCountRequest(request, response, filter, mimeType, zip);
 	}
